feat: add spinner
