+++ src/App.js
@@ -1,15 +1,14 @@
-import Footer from './components/Footer';
-import Header from './components/Header';
-import Main from './components/Main-new';
+import Footer from "./components/Footer";
+import Header from "./components/Header";
+import Main from "./components/Main";
 
 function App() {
   return (
-   <>
-        <Header />
-        <Main />
-        <Footer />
- </>
+    <>
+      <Header />
+      <Main />
+      <Footer />
+    </>
   );
 }
-
 export default App;
