+++ src/index.js
@@ -1,13 +1,7 @@
 import 'bootstrap/dist/css/bootstrap.min.css';
-import $ from 'jquery';
-import Popper from 'popper.js';
-import 'bootstrap/dist/js/bootstrap.bundle.min';
-
 import React from 'react';
 import ReactDOM from 'react-dom';
-import './index.css';
 import App from './App';
-
 
 ReactDOM.render(
   <React.StrictMode>
