+++ src/styles/style.css
@@ -1,10 +1,6 @@
 main {
   display: flex;
-  justify-content: center;
-}
-
-.main_wrapper {
-  display: flex;
+  align-items: center;
   flex-direction: column;
 }
 
@@ -14,7 +10,8 @@ main {
 .new_game {
   display: flex;
   flex-direction: column;
-  margin-top: 10%;
+  align-items: center;
+  margin-top: 5%;
 }
 
 .score {
